pyproject.toml
[project]
name = "iot-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mediapipe>=0.10.21",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "scikit-learn>=1.6.1",
    "tensorflow>=2.18.1",
    "tensorflow-io-gcs-filesystem==0.31.0",
    "websockets==15.0.1",
]

[tool.uv.workspace]
